refactor:additional code cleanup
