+++ src/britive/helpers/__init__.py
@@ -1,8 +1,6 @@
-from .custom_attributes import CustomAttributes
 from .methods import HelperMethods
 
 
 class Helpers:
     def __init__(self, britive) -> None:
-        self.custom_attributes = CustomAttributes(britive)
         self.helper_methods = HelperMethods(britive)
